video-learning-helper-backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
supabase==2.0.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jwt==3.3.4
bcrypt==4.1.1

# 任务队列
celery==5.3.4
redis==5.0.1

# 视频处理 - 使用更新的版本
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
whisper-openai==20231117
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
Pillow==10.1.0
moviepy==1.0.3
librosa==0.10.1
soundfile==0.12.1

# 音频处理与转写
openai-whisper==20231117

# PDF生成
reportlab==4.0.7
fpdf2==2.7.6

# 文件处理
aiofiles==23.2.1

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
email-validator==2.1.0
pydantic[email]==2.5.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 额外工具
imageio==2.33.0
imageio-ffmpeg==0.4.9
tqdm==4.66.1 